Neap Box 1.2.0

+++ bootstrap/cleanup.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+. /vagrant/resources/colors.sh
+
+# This script needs admin rights
+if [ 0 != $(id -u) ]; then
+	echox "${text_red}Error:${text_reset} This script must be run as root!"
+	exit 1
+fi
+
 # Removing leftover leases and persistent rules
 echo "Cleaning up dhcp leases and rules"
 rm /var/lib/dhcp/*
@@ -24,6 +32,16 @@ apt-get -y -q dist-upgrade
 echo "Clean packages"
 apt-get -y -q autoremove
 apt-get -y -q clean
+apt-get -y -q autoclean
+
+echo "Remove APT related files"
+find /var/lib/apt -type f | xargs rm -f
+
+echo "Remove unnecessary documentation"
+find /var/lib/doc -type f | xargs rm -f
+
+echo "Cleanup log files"
+find /var/log -type f | while read f; do echo -ne '' > $f; done;
 
 echo "Remove all temporary files"
 rm -rf /tmp/*
@@ -31,6 +49,11 @@ rm -rf /usr/src/*
 rm -rf /var/tmp/*
 rm -rf /var/log/*.log
 rm -rf /var/log/**/*.log
+
+echo "Clean history"
+unset histfile
+rm -f /root/.bash_history
+rm -f /home/vagrant/.bash_history
 
 # To remove history, execute the following logged in through SSH:
 # cat /dev/null > ~/.bash_history && history -c && exit
